+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added Classification.point_in_polygon?/3
 - Added Classification.points_within_polygon/2
 - Added Classification.nearest_point/3
+- Added Measure.centroid/1
 
 ## 0.3.1